+++ test/fixtures/done-and-promise.js
@@ -1,0 +1,6 @@
+When(async function (done) {
+  done();
+});
+Then(function () {
+  return true;
+});