+++ test/fixtures/given-rejects.js
@@ -1,0 +1,4 @@
+Given('x', () => Promise.reject(new Error('given blew up')));
+Then(function () {
+  return this.x === 1;
+});