+++ test/fixtures/resolves-false.js
@@ -1,0 +1,3 @@
+Then(function () {
+  return Promise.resolve(false);
+});