+++ test/fixtures/when-rejects.js
@@ -1,0 +1,6 @@
+When(function () {
+  return Promise.reject(new Error('when blew up'));
+});
+Then(function () {
+  return true;
+});